Add/View Users
Defines namespace and class named UserController to view the user name stored in the database. Class UserController also has the function to store and redirect the user to the path specified. It uses the function redirect (defined in core/bootsrap.php) to redirect the user to specified location, in this case its user page.